Mv copy_to_system from blivet.util to blivet.osinstall
As this seemed installer-specific, it seemed fine to move it out of
blivet.util and into blivet.osinstall.
